Updated download script

+++ ..Update-Modpack.ps1
@@ -1,30 +1,70 @@
-$url = "https://github.com/Clysop/ClyPack-1.16.4/archive/master.zip"
-$output = "modpack"
+param (
+  [switch]$SkipConfirm,
+  [switch]$SkipDownload
+)
 
-$mmc_file = "..\mmc-pack.json"
-$mi_file = "minecraftinstance.json"
+$REPO = 'Clysop/ClyPack-1.16.4'
+$BRANCH = 'master'
 
-cls
-Write-Output "Setting up modpack...`n`n`n`n`n`n"
-Write-Output "Downloading modpack..."
+$output = "$PSScriptRoot\modpack"
 
-Invoke-WebRequest -Uri $url -OutFile ($output + ".zip")
+$mmc_file = "$PSScriptRoot\..\mmc-pack.json"
+$mi_file = "$PSScriptRoot\minecraftinstance.json"
 
-Write-Output "Download complete."
-Write-Output "Extracting modpack..."
+$archive_url = "https://github.com/$REPO/archive/$BRANCH.zip"
+$api_url = "https://api.github.com/repos/$REPO/branches/$BRANCH"
 
-Expand-Archive ($output + ".zip") -Force
-Copy-Item ($output + "\*\*") "." -Force -Recurse
-Remove-Item $output -Recurse
-Remove-Item ($output + ".zip")
 
-Write-Output "Extraction complete."
+if (!($SkipDownload.IsPresent)) {
+  cls
+  Write-Output "Setting up modpack...`n`n`n`n`n`n"
+
+  # Check if there is a newer version available
+  $api_response = Invoke-RestMethod $api_url
+  $remote_date = Get-Date $api_response.commit.commit.author.date
+  if ((Test-Path $mi_file) -and ((Get-Item $mi_file).LastWriteTime -gt $remote_date)) {
+    if ($SkipConfirm.IsPresent) {
+      Exit
+    }
+    $confirmation = Read-Host "There doesn't seem to be a new version available, do you want to continue anyway? [Y/N]"
+    if ($confirmation.ToLower() -ne "y") {
+      Exit
+    }
+  }
+
+  $script_hash = (Get-FileHash $PSCommandPath).hash
+
+  Write-Output "Downloading modpack..."
+
+  Invoke-WebRequest $archive_url -OutFile ("$output.zip")
+
+  Write-Output "Download complete."
+  Write-Output "Extracting modpack..."
+
+  Expand-Archive ($output + ".zip") -Force
+  Copy-Item ($output + "\*\*") "." -Force -Recurse
+  Remove-Item $output -Recurse
+  Remove-Item ($output + ".zip")
+  
+  (Get-Item $mi_file).LastWriteTime = Get-Date
+  
+  Write-Output "Extraction complete."
+
+  if ($script_hash -ne (Get-FileHash $PSCommandPath).hash) {
+    Write-Output "`nUpdate script changed, reloading script.`n"
+    Invoke-Expression "$PSCommandPath -SkipDownload"
+    Exit
+  }
+}
+
 Write-Output "Downloading mods...`n"
 
 java -jar InstanceSync.jar
 
-Move-Item "thirdparty-mods\*" "mods" -Force
-Remove-Item "thirdparty-mods" -Recurse
+if (Test-Path "$PSScriptRoot\thirdparty-mods") {
+  Move-Item "thirdparty-mods\*" "mods" -Force
+  Remove-Item "thirdparty-mods" -Recurse
+}
 
 if (Test-Path $mmc_file) {
   $mi_content = Get-Content $mi_file | ConvertFrom-Json
